LayoutTests/inspector/debugger/async-stack-trace-expected.txt

Tests for async stack traces.


== Running test suite: AsyncStackTrace
-- Running test case: CheckAsyncStackTrace.RequestAnimationFrame
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
-- [N] requestAnimationFrame ----
1: [F] testRequestAnimationFrame
2: [P] Global Code

-- Running test case: CheckAsyncStackTrace.SetTimeout
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
-- [N] setTimeout ----
1: [F] testSetTimeout
2: [P] Global Code

-- Running test case: CheckAsyncStackTrace.SetInterval
PAUSE #1
CALL STACK:
0: [F] intervalFired
-- [N] setInterval ----
1: [F] testSetInterval
2: [P] Global Code
PAUSE #2
CALL STACK:
0: [F] intervalFired
-- [N] setInterval ----
1: [F] testSetInterval
2: [P] Global Code
PAUSE #3
CALL STACK:
0: [F] intervalFired
-- [N] setInterval ----
1: [F] testSetInterval
2: [P] Global Code

-- Running test case: CheckAsyncStackTrace.ChainedRequestAnimationFrame
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
-- [N] requestAnimationFrame ----
1: [F] testRequestAnimationFrame
-- [N] requestAnimationFrame ----
2: [F] testChainedRequestAnimationFrame
3: [P] Global Code

-- Running test case: CheckAsyncStackTrace.ReferenceCounting
PAUSE #1
CALL STACK:
0: [F] pauseThenFinishTest
-- [N] setTimeout ----
1: [F] intervalFired
-- [N] setInterval ----
2: [F] testReferenceCounting
3: [P] Global Code
-- Running test setup.
Save DebuggerManager.asyncStackTraceDepth

-- Running test case: AsyncStackTrace.DisableStackTrace
PASS: Async stack trace should be null.
-- Running test teardown.
Restore DebuggerManager.asyncStackTraceDepth
-- Running test setup.
Save DebuggerManager.asyncStackTraceDepth

-- Running test case: AsyncStackTrace.SetStackTraceDepth
PASS: Number of call frames should be equal to the depth setting.
-- Running test teardown.
Restore DebuggerManager.asyncStackTraceDepth
